Allow prometheus client to handle invalid UTF8
We have an ongoing issue where invalid UTF8 is being passed to the
prometheus client which blows up preventing us serving metrics. We have
tried to resolve this in fluentd but have been unable to so as a stop gap
(a possibly long one) we are patching the ruby prometheus client to
siliently handle invalid UTF8 and therefore continute to serve metrics.